Refactored(query-sort.js) seperated into query-group-by, query-order-by and query-limit components
